modularise 3 routes modularised welcome, my posts and write posts
