+++ database/db_connection.js
@@ -8,7 +8,7 @@ if (process.env.ENV==='test') {
   environment('config.env');
 }
 
-console.log(process.env.ENV);
+console.log('environment', process.env.ENV);
 
 if (!process.env.DATABASE_URL) throw new Error('Environment variable DATABASE_URL must be set');
