+++ src/routes.js
@@ -1,0 +1,6 @@
+module.exports = [
+  require('./routes/welcome.js'),
+  require('./routes/write_post.js'),
+  require('./routes/my_posts.js'),
+];
+
