+++ src/routes/write_post.js
@@ -1,0 +1,7 @@
+module.exports = {
+  method: 'GET',
+  path: '/write-post',
+  handler: {
+    view: 'write-post'
+  }
+};
