+++ src/server.js
@@ -11,6 +11,7 @@ const request = require('request');
 const fs = require('fs');
 const jwt = require('jsonwebtoken');
 const hapiJwt = require('hapi-auth-jwt2');
+const routes = require('./routes.js');
 
 const server = new hapi.Server();
 
@@ -64,13 +65,6 @@ server.register([inert, credentials, vision, CookieAuth], (err) => {
   });
 
 
-  server.route({
-    method: 'GET',
-    path: '/write-post',
-    handler: {
-      view: 'write-post'
-    }
-  });
 
   server.route({
     method: 'GET',
@@ -83,62 +77,8 @@ server.register([inert, credentials, vision, CookieAuth], (err) => {
   });
 
 
-  server.route({
-    method: 'GET',
-    path: '/welcome',
-    handler: (req, reply) => {
-      const query = req.url.query;
-      const gitHubUrl = `https://github.com/login/oauth/access_token?client_id=${process.env.CLIENT_ID}&client_secret=${process.env.CLIENT_SECRET}&code=${query.code}`;
+  server.route(routes);
 
-        request.post(gitHubUrl, (err, res, body) => {
-
-        const accessToken = querystring.parse(body).access_token;
-        const headers = {
-          'User-Agent': 'oauth_github_jwt',
-          Authorization: `token ${accessToken}`
-        };
-        request.get({url: 'https://api.github.com/user', headers}, (err, res, body) => {
-          const parsedBody = JSON.parse(body);
-        const userData = {
-          'username': parsedBody.login,
-          'avatar': parsedBody.avatar_url,
-          'userId': parsedBody.id,
-          accessToken
-        };
-        req.cookieAuth.set({ avatarUrl:userData.avatar, username: userData.username, accessToken: userData.accessToken });
-        postData.checkUser(userData, (dbErr, dbRes) => {
-          console.log(dbErr);
-          data.getBlogPosts((dbErr, res) => {
-            if (dbErr) {
-              reply.view({ message:'Lo sentimos, actualmente estamos experimentando dificultades con el servidor' });
-              return;
-            }
-            reply.redirect('/');
-            // reply.view('index', { 
-            //   res
-            // });
-          });
-        });
-        });
-      });
-    }
-  });
-
-
-  server.route({
-    method: 'GET',
-    path: '/my-posts',
-    handler: (req, reply) => {
-      data.getBlogPostsByUser(req.auth.credentials.username, (dbErr, res) => {
-        if (dbErr) {
-          reply.view(index, { message: 'Lo sentimos, actualmente estamos experimentando dificultades con el servidor' });
-          return;
-        }
-        console.log('res', res);
-        reply.view('index', { res });
-      });
-    }
-  });
 
   server.route({
     method: 'POST',
